experiments/regex/requirements.txt
google-re2==1.1.20240702
numpy==2.1.3
matplotlib==3.9.2